Move Project and corresponding utils on board of libmmv_model library
