+++ examples/quickrefl/core/editors/welcomeview/CMakeLists.txt
@@ -1,15 +1,9 @@
 set(library_name quickreflcore)
 
 target_sources(${library_name} PRIVATE
-    modelhaschangedcontroller.cpp
-    modelhaschangedcontroller.h
     openprojectwidget.cpp
     openprojectwidget.h
-    project.cpp
-    project.h
     project_types.h
-    projectchangecontroller.cpp
-    projectchangecontroller.h
     projectmanager.cpp
     projectmanager.h
     projectmanagerdecorator.cpp
@@ -18,8 +12,6 @@ target_sources(${library_name} PRIVATE
     projectmanagerinteractor.h
     projectpanewidget.cpp
     projectpanewidget.h
-    projectutils.cpp
-    projectutils.h
     recentprojectwidget.cpp
     recentprojectwidget.h
     welcomeview.cpp
