+++ examples/quickrefl/core/editors/welcomeview/welcomeview.cpp
@@ -12,7 +12,7 @@
 #include "openprojectwidget.h"
 #include "projectmanagerdecorator.h"
 #include "projectmanagerinteractor.h"
-#include "projectutils.h"
+#include <mvvm/project/projectutils.h>
 #include "recentprojectwidget.h"
 #include "welcomeviewsettings.h"
 #include <QApplication>
