+++ examples/quickrefl/tests/modelhaschangedcontroller.test.cpp
@@ -8,7 +8,7 @@
 // ************************************************************************** //
 
 #include "google_test.h"
-#include "modelhaschangedcontroller.h"
+#include <mvvm/project/modelhaschangedcontroller.h>
 #include <mvvm/model/propertyitem.h>
 #include <mvvm/model/sessionmodel.h>
