+++ examples/quickrefl/tests/projectchangecontroller.test.cpp
@@ -8,7 +8,7 @@
 // ************************************************************************** //
 
 #include "google_test.h"
-#include "projectchangecontroller.h"
+#include <mvvm/project/projectchangecontroller.h>
 #include <mvvm/model/propertyitem.h>
 #include <mvvm/model/sessionmodel.h>
