+++ source/libmvvm_model/mvvm/project/CMakeLists.txt
@@ -1,2 +1,10 @@
 target_sources(mvvm_model PRIVATE
+    modelhaschangedcontroller.cpp
+    modelhaschangedcontroller.h
+    project.cpp
+    project.h
+    projectchangecontroller.cpp
+    projectchangecontroller.h
+    projectutils.cpp
+    projectutils.h
 )
