+++ source/libmvvm_model/mvvm/project/projectchangecontroller.cpp
@@ -8,7 +8,7 @@
 // ************************************************************************** //
 
 #include "projectchangecontroller.h"
-#include "modelhaschangedcontroller.h"
+#include <mvvm/project/modelhaschangedcontroller.h>
 #include <mvvm/model/sessionmodel.h>
 
 using namespace ModelView;
