+++ source/libmvvm_model/mvvm/project/projectutils.cpp
